Tmux migration

+++ bin/omadeb-sub/theme.sh
@@ -1,12 +1,17 @@
 #!/bin/bash
 
-THEME_NAMES=("Tokyo Night" "Catppuccin" "Nord" "Everforest" "Gruvbox" "Kanagawa" "Ristretto" "Rose Pine" "Matte Black" "Osaka Jade")
+THEME_NAMES=("Tokyo Night" "Catppuccin" "Catppuccin Custom" "Nord" "Everforest" "Gruvbox" "Kanagawa" "Ristretto" "Rose Pine" "Matte Black" "Osaka Jade")
 THEME=$(gum choose "${THEME_NAMES[@]}" "<< Back" --header "Choose your theme" --height 12 | tr '[:upper:]' '[:lower:]' | sed 's/ /-/g')
 
 if [ -n "$THEME" ] && [ "$THEME" != "<<-back" ]; then
-  cp $OMADEB_PATH/themes/$THEME/alacritty.toml ~/.config/alacritty/theme.toml
-  cp $OMADEB_PATH/themes/$THEME/zellij.kdl ~/.config/zellij/themes/$THEME.kdl
-  sed -i "s/theme \".*\"/theme \"$THEME\"/g" ~/.config/zellij/config.kdl
+  if [ -f ~/.config/alacritty/theme.toml ]; then
+    cp $OMADEB_PATH/themes/$THEME/alacritty.toml ~/.config/alacritty/theme.toml
+  fi
+
+  if [ -f ~/.config/zellij/themes/$THEME.kdl ]; then
+    cp $OMADEB_PATH/themes/$THEME/zellij.kdl ~/.config/zellij/themes/$THEME.kdl
+    sed -i "s/theme \".*\"/theme \"$THEME\"/g" ~/.config/zellij/config.kdl
+  fi
   if [ -d "$HOME/.config/nvim" ]; then
     cp $OMADEB_PATH/themes/$THEME/neovim.lua ~/.config/nvim/lua/plugins/theme.lua
   fi
